+++ src/common/gateway/entities/tx.in.pool.response.ts
@@ -1,0 +1,9 @@
+import { TxInPoolFields } from "./tx.in.pool.fields";
+
+export class TxInPoolResponse {
+  constructor(init?: Partial<TxInPoolResponse>) {
+    Object.assign(this, init);
+  }
+
+  txFields: TxInPoolFields = new TxInPoolFields();
+}
