+++ src/common/gateway/entities/tx.pool.ts
@@ -1,0 +1,11 @@
+import { TxInPoolResponse } from "./tx.in.pool.response";
+
+export class TxPool {
+  constructor(init?: Partial<TxPool>) {
+    Object.assign(this, init);
+  }
+
+  regularTransactions: TxInPoolResponse[] = [];
+  smartContractResults: TxInPoolResponse[] = [];
+  rewards: TxInPoolResponse[] = [];
+}
